Wrap state in `Arc` to avoid cloning it.
axum clones state on every request, making it very expensive. We can
avoid this by wrapping it in an `Arc`, which means we just clone the
reference.

As state is mostly read-only, this seems to work without further
changes, except for explicitly cloning the metrics registry too. (The
metrics registry is itself cheap to clone, so no issues there.)