Move configuration when we don't need to hold onto it.
There are certain functions in the `Connector` trait that use the raw
configuration, and accept a reference. We don't actually need to keep
the configuration around in the SDK so we can pass by value here, which
means that the relevant functions can own the data if they require it.

Making these input parameters values rather than references allows us to
avoid cloning in other locations later.