Introduce a "fast path" for returning already-serialized JSON.
This introduces a new enum, `JsonResponse`, which is designed to
supplant `axum::Json`. It implements `axum::IntoResponse`.

It has two variants:

1. `Value`, which behaves the same as `axum::Json`.
2. `Serialized`, which allows the connector to provide
   already-serialized JSON in the form of a `Bytes` value.

The latter can be used to construct the JSON directly in the database
and avoid deserialization and re-serialization.

This is a breaking change, as at the very least, the connector
implementations will now need to wrap responses in
`JsonResponse::Value`. Hopefully it's not too much of a big deal.

There is a helper function, `JsonResponse::into_value`, used to
deserialize the serialized bytes in certain situations. This is because
the v2 compatibility layer and the connector used for ndc-test both
require transforming the data, and so cannot work directly with bytes.
We should not call this when running the connector normally, and it is
not available outside this crate.