Allow the operator to set the listening host IP address. (#122)
This is often useful for security reasons, e.g. only listening on
localhost or an internal IP when the service is to be proxied.

It is also useful in testing on macOS, as running the service bound to a
localhost IP (`127.0.0.1` or `::1`) means that the firewall will not be
alerted, and so the test will not trigger an allow/deny prompt.