fix tags for registry connectors (#140)
Tags seem to be an enum in our current system that allows only the
following values:


![image](https://github.com/hasura/ndc-hub/assets/4211715/0cb85a4c-edd4-4fe4-89b4-7c3068ec3bad)

We already started using new registry Hub API internally and we will
need to soon migrate other systems to use those APIs. There is no
concept of tags in the new hub API. When we introduce, we can overcome
the limitation of having it as enum and instead allow it to be array of
any string.