Docs: Update READMEs
Updates READMEs to include the new `ddn connector init -i` workflow. Additionally, it standarizes the READMEs by reducing the amount of updates needed to CLI commands and — instead — links out to the docs. For any README that is of a "special child" (e.g., PostgreSQL), we've linked to the docs page(s).