RFC: Include CLI packaging information within the connector metadata  (#322)
# Add Binary CLI Plugin Packaging Information to
`connector-metadata.yaml`

This PR introduces changes to improve the management of binary CLI
plugin information in the `connector-metadata.yaml` file.

## Key Changes

1. Extended `connector-metadata.json` to include binary CLI plugin
packaging information.
2. Added a new optional `version` field to the
`ConnectorMetadataDefinition` type for future versioning.
3. Extend `BinaryCliPluginDefinition` to accommodate inline binary CLI
definition.

## Addition of `connector-metadata-types`

A significant part of this PR is the introduction of
`connector-metadata-types`. This file contains TypeScript type
definitions for the `connector-metadata.json` structure and is meant to
act as the source of truth for the schema of the
`connector-metadata.json` file. As we advance, it is expected that if
any changes are proposed in the connector metadata structure, they also
update the `connectorTypes.ts` file in the `connector-metadata-types`
folder.