fix broken links in mongodb connector page (#392)
The documentation link on the MongoDB connector page is broken, so I
updated it.

The prerequisites sections has broken links to steps on creating a
supergraph and a subgraph. The docs are structured differently now so I
didn't see clear pages to switch those links to. Plus the connector
documentation link now covers all of the prerequisite steps so I thought
it would be best to remove the prerequisites section from the hub.

I also switch the issues link to point to the connector repo instead of
the engine repo.

---------

Co-authored-by: Rob Dominguez <rob.dominguez@hasura.io>